moved page context to separate file to resolve invalid page export field error
