+++ contexts/FunkContext.tsx
@@ -1,0 +1,7 @@
+"use client";
+import { createContext } from "react";
+
+export const FunkContext = createContext({
+  isFunkEnabled: false,
+  setIsFunkEnabled: (enabled: boolean) => {}
+});
